Created LoginController

+++ PetBaseData.API/Entities/PetObject.cs
@@ -14,7 +14,7 @@ namespace PetBaseData.API.Entities
         public string Id { get; set; }
         public string PetName { get; set; }
         public int Ability { get; set; }
-        public float AbilityMulti { get; set; }
+        public double AbilityMulti { get; set; }
         public string PetGrade { get; set; }
         public int Level { get; set; }
         public int MaxStar { get; set; }
